Add naive migration from 1:m to m:n relation; data loss!
This is an intentionally bad example. If you were to run this migration,
which was naively generated by Django with `./manage.py makemigrations',
Django would _drop_ the `label_id` field on `tasks` and create a new
`tasks_labels` table... losing all the Label associations that already
exist in the database.

This is an example of a schema migration that is in appropriately
generated by Django. Instead of dropping the field, creating the
table, and losing the data, you should create the table, _migrate the
data_ by creating new records in `tasks_labels`, and _then_ drop the
old `label_id` field. Django can't figure that out for you, though.

See https://docs.djangoproject.com/en/1.9/topics/migrations/#data-migrations